pull

+++ indra/llcommon/llqueuedthread.cpp
@@ -1,31 +1,25 @@
 /** 
  * @file llqueuedthread.cpp
  *
- * $LicenseInfo:firstyear=2004&license=viewergpl$
- * 
- * Copyright (c) 2004-2009, Linden Research, Inc.
- * 
+ * $LicenseInfo:firstyear=2004&license=viewerlgpl$
  * Second Life Viewer Source Code
- * The source code in this file ("Source Code") is provided by Linden Lab
- * to you under the terms of the GNU General Public License, version 2.0
- * ("GPL"), unless you have obtained a separate licensing agreement
- * ("Other License"), formally executed by you and Linden Lab.  Terms of
- * the GPL can be found in doc/GPL-license.txt in this distribution, or
- * online at http://secondlifegrid.net/programs/open_source/licensing/gplv2
+ * Copyright (C) 2010, Linden Research, Inc.
  * 
- * There are special exceptions to the terms and conditions of the GPL as
- * it is applied to this Source Code. View the full text of the exception
- * in the file doc/FLOSS-exception.txt in this software distribution, or
- * online at
- * http://secondlifegrid.net/programs/open_source/licensing/flossexception
+ * This library is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation;
+ * version 2.1 of the License only.
  * 
- * By copying, modifying or distributing this software, you acknowledge
- * that you have read and understood your obligations described above,
- * and agree to abide by those obligations.
+ * This library is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * Lesser General Public License for more details.
  * 
- * ALL LINDEN LAB SOURCE CODE IS PROVIDED "AS IS." LINDEN LAB MAKES NO
- * WARRANTIES, EXPRESS, IMPLIED OR OTHERWISE, REGARDING ITS ACCURACY,
- * COMPLETENESS OR PERFORMANCE.
+ * You should have received a copy of the GNU Lesser General Public
+ * License along with this library; if not, write to the Free Software
+ * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
+ * 
+ * Linden Research, Inc., 945 Battery Street, San Francisco, CA  94111  USA
  * $/LicenseInfo$
  */
 
@@ -140,8 +134,8 @@ S32 LLQueuedThread::updateQueue(F32 max_time_ms)
 		pending = getPending();
 		if(pending > 0)
 		{
-		unpause();
-	}
+			unpause();
+		}
 	}
 	else
 	{
@@ -251,7 +245,7 @@ bool LLQueuedThread::addRequest(QueuedRequest* req)
 // MAIN thread
 bool LLQueuedThread::waitForResult(LLQueuedThread::handle_t handle, bool auto_complete)
 {
-	llassert (handle != nullHandle())
+	llassert (handle != nullHandle());
 	bool res = false;
 	bool waspaused = isPaused();
 	bool done = false;
@@ -264,7 +258,7 @@ bool LLQueuedThread::waitForResult(LLQueuedThread::handle_t handle, bool auto_co
 		{
 			done = true; // request does not exist
 		}
-		else if (req->getStatus() == STATUS_COMPLETE)
+		else if (req->getStatus() == STATUS_COMPLETE && !(req->getFlags() & FLAG_LOCKED))
 		{
 			res = true;
 			if (auto_complete)
@@ -372,9 +366,17 @@ bool LLQueuedThread::completeRequest(handle_t handle)
 #if _DEBUG
 // 		llinfos << llformat("LLQueuedThread::Completed req [%08d]",handle) << llendl;
 #endif
-		mRequestHash.erase(handle);
-		req->deleteRequest();
-// 		check();
+		if (!(req->getFlags() & FLAG_LOCKED))
+		{
+			mRequestHash.erase(handle);
+			req->deleteRequest();
+// 			check();
+ 		}
+ 		else
+ 		{
+			// Cause deletion immediately after FLAG_LOCKED is released.
+ 			req->setFlags(FLAG_AUTO_COMPLETE);
+ 		}
 		res = true;
 	}
 	unlockData();
@@ -421,12 +423,44 @@ S32 LLQueuedThread::processNextRequest()
 		if ((req->getFlags() & FLAG_ABORT) || (mStatus == QUITTING))
 		{
 			req->setStatus(STATUS_ABORTED);
+			// Unlock, because we can't call finishRequest() while keeping this lock:
+			// generateHandle() takes this lock too and is called while holding a lock
+			// (ie LLTextureFetchWorker::mWorkMutex) that finishRequest will lock too,
+			// causing a dead lock.
+			// Although a complete rewrite of LLQueuedThread is in order because it's
+			// far from robust the way it handles it's locking; the following rationale
+			// at least makes plausible that releasing the lock here SHOULD work if
+			// the original coder didn't completely fuck up: if before the QueuedRequest
+			// req was only accessed while keeping the lock, then it still should
+			// never happen that another thread is waiting for this lock in order to
+			// access the QueuedRequest: a few lines lower we delete it.
+			// In other words, if directly after releasing this lock another thread
+			// would access req, then that can only happen by finding it again in
+			// either mRequestQueue or mRequestHash. We already deleted it from the
+			// first, so this access would have to happen by finding it in mRequestHash.
+			// Such access happens in the following functions:
+			// 1) LLQueuedThread::shutdown -- but it does that anyway, as it doesn't use any locking.
+			// 2) LLQueuedThread::generateHandle -- might skip our handle while before it would block until we deleted it. Skipping it is actually better.
+			// 3) LLQueuedThread::waitForResult -- this now doesn't touch the req when it has the flag FLAG_LOCKED set.
+			// 4) LLQueuedThread::getRequest -- whereever this is used, FLAG_LOCKED is tested before using the req.
+			// 5) LLQueuedThread::getRequestStatus -- this is a read-only operation on the status, which should never be changed from finishRequest().
+			// 6) LLQueuedThread::abortRequest -- it doesn't seem to hurt to add flags (if this happens at all), while calling finishRequest().
+			// 7) LLQueuedThread::setFlags -- same.
+			// 8) LLQueuedThread::setPriority -- doesn't access req with status STATUS_ABORTED, STATUS_COMPLETE or STATUS_INPROGRESS.
+			// 9) LLQueuedThread::completeRequest -- now sets FLAG_AUTO_COMPLETE instead of deleting the req, if FLAG_LOCKED is set, so that deletion happens here when finishRequest returns.
+			req->setFlags(FLAG_LOCKED);
+			unlockData();
 			req->finishRequest(false);
-			if (req->getFlags() & FLAG_AUTO_COMPLETE)
+			lockData();
+			req->resetFlags(FLAG_LOCKED);
+			if ((req->getFlags() & FLAG_AUTO_COMPLETE))
 			{
+				req->resetFlags(FLAG_AUTO_COMPLETE);
 				mRequestHash.erase(req);
-				req->deleteRequest();
 // 				check();
+				unlockData();
+				req->deleteRequest();
+				lockData();
 			}
 			continue;
 		}
@@ -453,14 +487,23 @@ S32 LLQueuedThread::processNextRequest()
 		{
 			lockData();
 			req->setStatus(STATUS_COMPLETE);
-			req->finishRequest(true);
-			if (req->getFlags() & FLAG_AUTO_COMPLETE)
-			{
-				mRequestHash.erase(req);
-				req->deleteRequest();
-// 				check();
-			}
+			req->setFlags(FLAG_LOCKED);
 			unlockData();
+			req->finishRequest(true);
+			if ((req->getFlags() & FLAG_AUTO_COMPLETE))
+			{
+				lockData();
+				req->resetFlags(FLAG_AUTO_COMPLETE);
+				mRequestHash.erase(req);
+// 				check();
+				req->resetFlags(FLAG_LOCKED);
+				unlockData();
+				req->deleteRequest();
+			}
+			else
+			{
+				req->resetFlags(FLAG_LOCKED);
+			}
 		}
 		else
 		{
